econ-file-factory/backend/requirements.txt
Flask
pandas
Werkzeug
flask-cors
openai>=1.0.0
numpy
openpyxl
redis
python-dotenv
chardet 